+++ ios/Classes/AudioWaveformsPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface AudioWaveformsPlugin : NSObject<FlutterPlugin>
-@end