+++ macos/audio_waveforms/Sources/audio_waveforms/AudioPlayer.swift
@@ -1,3 +1,4 @@
+import FlutterMacOS
 import AVKit
 import Foundation
 
@@ -10,7 +11,7 @@ class AudioPlayer: NSObject, AVAudioPlayerDelegate {
     private var player: AVAudioPlayer?
     private var finishMode: FinishMode = FinishMode.stop
     private var updateFrequency = 200
-    var plugin: SwiftAudioWaveformsPlugin
+    var plugin: AudioWaveformsPlugin
     var playerKey: String
     var flutterChannel: FlutterMethodChannel
 
@@ -20,7 +21,7 @@ class AudioPlayer: NSObject, AVAudioPlayerDelegate {
     ///   - playerKey: Unique identifier for this player
     ///   - channel: Flutter method channel for callbacks
     init(
-        plugin: SwiftAudioWaveformsPlugin, playerKey: String,
+        plugin: AudioWaveformsPlugin, playerKey: String,
         channel: FlutterMethodChannel
     ) {
         self.plugin = plugin